unify modals in kobo settings modal
