--- apps/back-office/src/app/components/synchronize-kobo-at-modal/synchronize-kobo-at-modal.component.html
@@ -1,60 +1,0 @@
-<ui-dialog size="medium" [closable]="false">
-  <ng-container ngProjectAs="header">
-    <h3 class="font-semibold">{{ 'components.form.create.kobo.data.synchronizeAt' | translate }}</h3>
-  </ng-container>
-
-  <ng-container ngProjectAs="content">
-    <ui-spinner *ngIf="loading" size="medium"></ui-spinner>
-
-    <div
-      class="flex flex-col items-center my-2"
-      *ngIf="formGroup.get('cronSchedule')?.value"
-    >
-      <ui-button category="secondary" variant="primary" (click)="onReset()">{{
-        'kendo.grid.columnsReset' | translate
-      }}</ui-button>
-    </div>
-
-    <ui-alert>
-      {{ 'components.form.create.kobo.data.fromFormWithVersion' | translate : { formName: data.form.name, versionId: data.deployedVersionId} }}
-    </ui-alert>
-
-    <form [formGroup]="formGroup" class="mt-2">
-      <ui-toggle formControlName="dataFromDeployedVersion" *ngIf="formGroup.get('cronSchedule')?.value">
-        <ng-container ngProjectAs="label">
-          {{ 'components.form.create.kobo.data.syncOnlyFromDeployedVersion' | translate }}
-          <ui-icon
-            class="ml-1 cursor-help self-center"
-            icon="info_outline"
-            variant="grey"
-            [size]="18"
-            [uiTooltip]="
-              'components.form.create.kobo.data.tooltip.syncOnlyFromDeployedVersion' | translate
-            "
-          ></ui-icon>
-        </ng-container>
-      </ui-toggle>
-
-      <shared-cron-expression-control
-        class="flex mt-2"
-        formControlName="cronSchedule"
-        [toolTip]="'components.form.create.kobo.data.tooltip.schedule' | translate"
-      ></shared-cron-expression-control>
-    </form>
-  </ng-container>
-
-  <ng-container ngProjectAs="actions">
-    <ui-button [uiDialogClose]="updated ? formGroup.value: null" variant="default">{{
-      'common.close' | translate
-    }}</ui-button>
-    <ui-button
-      category="secondary"
-      variant="primary"
-      (click)="onSave()"
-      cdkFocusInitial
-      [disabled]="formGroup.pristine"
-    >
-      {{ 'common.save' | translate }}
-    </ui-button>
-  </ng-container>
-</ui-dialog>
