--- apps/back-office/src/app/components/synchronize-kobo-at-modal/synchronize-kobo-at-modal.component.spec.ts
@@ -1,22 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { SynchronizeKoboAtModalComponent } from './synchronize-kobo-at-modal.component';
-
-describe('SynchronizeKoboAtModalComponent', () => {
-  let component: SynchronizeKoboAtModalComponent;
-  let fixture: ComponentFixture<SynchronizeKoboAtModalComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [SynchronizeKoboAtModalComponent],
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(SynchronizeKoboAtModalComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
