--- apps/back-office/src/app/components/synchronize-kobo-modal/synchronize-kobo-modal.component.html
@@ -1,47 +1,0 @@
-<ui-dialog size="medium" [closable]="false">
-  <ng-container ngProjectAs="header">
-    <h3 class="font-semibold">{{ 'components.form.create.kobo.data.synchronize' | translate }}</h3>
-  </ng-container>
-  <ng-container ngProjectAs="content">
-    <ui-spinner *ngIf="loading" size="medium"></ui-spinner>
-    <div class="flex flex-col items-center my-2">
-      <ui-button category="secondary" variant="primary" (click)="onSynchronize()">{{
-        'components.form.create.kobo.data.start' | translate
-      }}</ui-button>
-    </div>
-    <ui-alert>
-      {{ 'components.form.create.kobo.data.fromFormWithVersion' | translate : { formName: data.form.name, versionId: data.deployedVersionId} }}
-    </ui-alert>
-    <form [formGroup]="formGroup" class="mt-2">
-      <ui-toggle formControlName="dataFromDeployedVersion" class="mb-5">
-        <ng-container ngProjectAs="label">
-          {{ 'components.form.create.kobo.data.syncOnlyFromDeployedVersion' | translate }}
-          <ui-icon
-            class="ml-1 cursor-help self-center"
-            icon="info_outline"
-            variant="grey"
-            [size]="18"
-            [uiTooltip]="
-              'components.form.create.kobo.data.tooltip.syncOnlyFromDeployedVersion' | translate
-            "
-          ></ui-icon>
-        </ng-container>
-      </ui-toggle>
-    </form>
-  </ng-container>
-
-  <ng-container ngProjectAs="actions">
-    <ui-button [uiDialogClose]="updated ? formGroup.get('dataFromDeployedVersion')?.value : null" variant="default">{{
-      'common.close' | translate
-    }}</ui-button>
-    <ui-button
-      category="secondary"
-      variant="primary"
-      (click)="onSave()"
-      cdkFocusInitial
-      [disabled]="formGroup.pristine"
-    >
-      {{ 'common.save' | translate }}
-    </ui-button>
-  </ng-container>
-</ui-dialog>
