+++ libs/shared/src/lib/components/controls/cron-expression-control/cron-expression-control.component.html
@@ -17,7 +17,7 @@
   />
   <ui-icon
     icon="edit"
-    class="cursor-pointer"
+    class="cursor-pointer grid"
     variant="grey"
     uiSuffix
     (click)="onEdit()"
